Add files via upload
Django signals do not run in the same thread as the caller by default.
This means that the signal handler function will be executed in a separate thread from the thread that triggered the event. This is done to prevent blocking the main thread and to allow for asynchronous processing.

In this example, the my_caller function creates and saves a MyModel object, which triggers the post_save signal. The my_handler function, which is the signal handler, will be executed in a separate thread. When you run this code, you'll notice that the thread IDs printed in the my_caller and my_handler functions are different, confirming that they are running in separate threads.