Question 3
Django signals do not run in the same database transaction as the caller by default.

This means that the signal handler function will be executed in a separate database transaction from the transaction that triggered the event. This is done to prevent potential issues if the signal handler were to modify the database and the original transaction were to fail.

In this example, the my_caller function creates a MyModel object within a transaction and then intentionally raises an exception to trigger a rollback. The my_handler function, which is the signal handler, will be executed in a separate transaction. If the signal handler were to modify the database and the original transaction were to fail, the changes made by the signal handler would not be rolled back.