Question 1
Django signals are executed synchronously by default.

This means that the signal handler function will be executed within the same request/response cycle as the event that triggered it.

In this example, whenever a MyModel object is saved, the my_handler function will be called immediately after the save operation is completed. This ensures that the signal handler's actions are part of the same transaction and can directly interact with the saved object.